+++ .gitignore
@@ -30,3 +30,6 @@
 *.exe
 *.out
 *.app
+
+# Build folder
+build
