+++ application/sources.cmake
@@ -1,0 +1,4 @@
+FILE(GLOB ${APP_LIBRARY}_SOURCES
+    "*.h"
+    "*.cc"
+)
