+++ cmake/include_directories.cmake
@@ -1,0 +1,3 @@
+SET(${CMAKE_PROJECT_NAME}_INCLUDE_DIRECTORIES
+    ${CMAKE_CURRENT_LIST_DIR}/..
+)
