+++ cmake/names.cmake
@@ -1,0 +1,3 @@
+SET(APP_LIBRARY  uurenderapp)
+SET(CMD_LIBRARY  uuappcommand)
+SET(MATH_LIBRARY uumath)
