+++ commands/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.12)
+project(${CMD_LIBRARY})
+
+include(sources.cmake)
+
+add_library(${CMD_LIBRARY} STATIC ${${CMD_LIBRARY}_SOURCES})
