+++ commands/open_command.cc
@@ -1,0 +1,10 @@
+//
+// Created by Иван Морозов on 15.11.2018.
+//
+
+#include "open_command.h"
+
+OpenCommand::OpenCommand(char *filepath) : _filepath(filepath) {}
+
+void OpenCommand::execute() {
+}
