+++ commands/open_command.h
@@ -5,15 +5,13 @@
 #ifndef KURSOVAYA2_0_OPEN_COMMAND_H
 #define KURSOVAYA2_0_OPEN_COMMAND_H
 
-#include "../command.h"
-#include "model.h"
+#include "command.h"
 
 class OpenCommand : public Command {
  public:
-  OpenCommand(Model **model, char *filepath);
+  OpenCommand(char *filepath);
   void execute() final;
  private:
-  Model **_model;
   char *_filepath;
 
 };
