+++ commands/sources.cmake
@@ -1,0 +1,4 @@
+FILE(GLOB ${CMD_LIBRARY}_SOURCES
+    "*.h"
+    "*.cc"
+)
