--- model/model.h
@@ -1,20 +1,0 @@
-#ifndef __MODEL_H__
-#define __MODEL_H__
-
-#include <vector>
-#include "geometry.h"
-
-class Model {
-private:
-	std::vector<Vec3f> verts_;
-	std::vector<std::vector<int> > faces_;
-public:
-	Model(const char *filename);
-	~Model();
-	int nverts();
-	int nfaces();
-	Vec3f vert(int i);
-	std::vector<int> face(int idx);
-};
-
-#endif //__MODEL_H__
