--- uml/Kursovaya.xml
@@ -1,1 +1,0 @@
-<mxfile userAgent="Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/11.1.2 Safari/605.1.15" version="9.4.2" editor="www.draw.io" type="device"><diagram id="a97d62c8-c010-d936-02a5-464a7a674fcd" name="Page-1">7Vpdb6M4FP01kdKVugqQkMxjQjq7D63Une5qZp4qF9zgrbEROE06v36vwXybDGk8lVZCihJ8uTb2Pef6HiATx4uOfyQoDu94gOnEngXHibOd2LZlWS78SMtbbllan3LDLiGBcqoMD+QHVsaZsu5JgNOGo+CcChI3jT5nDPuiYUNJwg9Nt2dOm1eN0Q53DA8+ol3rVxKIMLeu7GVl/xOTXVhc2XLV+iJUOKuVpCEK+KFmcm4mjpdwLvKj6OhhKoNXxCXv97nnbDmxBDMxpIOdd3hFdK/WdocI+0qYnFU+Q/FWLDs9kIgiBq3NM2fiQZ2xoI0o2TE49uG6OAHDK04EgYit1QnBY7D6IaHBLXrjezm7VCD/pWhtQp6QHzAsompMOJ0IBb7tNjweZE8wz8Ca4BR87oslWy3THTo2HG9RKpTB55SiOCVP5TIilOwI23AheKScVIBgOfjYG2SrhA44j3mERfIGLqrDXIGt2D5XQT9U1LFnyiWs08ZRRqTouitHrhCFAwWqHmCnA/A1NPcEvibOGr7/IXXEZ79NINS2c0SwdOmJ4rh0XccxBUwF4azl2aEKBEtkECb8BXuccuDElvGcO4TSlqmgD8XPopc8aYx8wna3mc92Xlm+qJhJE4e+zzRLqZAEAWYSeC6QQDnKEtKYEyaymC428IEoe7PfF5MFTNyDtlW14SPdE+FxBmuBQMnrYKDQAUsaDWOHrWeHooPtDqPDfH45G+YdNnSQoyRDJEeu2N2sd8EWAQAUVzj9LWHcXlsdLJ0ulo4GN4qeML3nKckY6GyT3LeF5wdAtpwNg2x1OWILDWKwhFmVtNO/AKUdFjIlr1rZy9ljvE/DzR42NPboQ/a+4GB6VWb0K4frjGlshBOrYZxY2JeTwu2QorE1j2XbSNl27J/X7bmubi8M1O2ltm4/+jyKEAvKBPbKdrtwP0ZKdueOhQgfq/ZQcrhnV20dGUxU7dVYtU1ApqvaOsgMVO1PPVUbynSWiLUSrE9M6ZxiodJ7WkvzqufEsyfrsYQbJIimhOsIsjRQwq2uFC9BHuu3kfrttup3Kcdq+FquDl8D9dvqFvA8q/ER+3uB+0X4dVXWp1djZg+lwvzccq2F3kS5drrlGcMNWpG3sNKQ7zhD9KayQvz2LMCBil4ooiId8ZGIb9IM0cpb35XTv1iIN5WoaC+4RKYc+5ZnZTsbgQVr+dATmk+UywSWps9ELiAbCFpqGKtKPTnl04kHK+T7xC+8VnoAEkzhtuS1OZYuvKrrvSROlcTLWSuJ28kJuxVUVdWrBVI5jWEp29VZXySr8bgxG34eumpi6hRPKeuFV5edlm1iZ+6Ks9P3S+O2ewLonqw/cZekA7bwuwTXghvjbdKFoBVPqH4GmoH7pGJ256uk3K2xP09rOevJizAx6idDnHA19846TrgGNmhn/jH66VxdNFRvvU8/9T1e/v/op2K/rb9MBqqP6smselotB6gn+1epp54XyqN6etcbpZ6cP6GedMAaUU/jq2EzoOnUkw40E+qp7+XwQPVU250z7dSQTt2DUUwZoohOTOkoYkRMzTpQGRBTxfH3SlgxmNi3ZvPcB1XvE06qIOU6RgXuQik1OLjOxwT31wZw2Q1g39PUSyMIzeq/jbnErP4h6tz8Bw==</diagram></mxfile>
